+++ lib/libsvm/core_ext.rb
@@ -1,0 +1,2 @@
+require 'libsvm/core_ext/array'
+require 'libsvm/core_ext/hash'