+++ lib/libsvm/core_ext/array.rb
@@ -1,0 +1,5 @@
+require 'libsvm/core_ext/collection'
+
+class Array
+  include Libsvm::CoreExtensions::Collection
+end