+++ lib/libsvm/core_ext/hash.rb
@@ -1,0 +1,5 @@
+require 'libsvm/core_ext/collection'
+
+class Hash
+  include Libsvm::CoreExtensions::Collection
+end